+++ test/reporter.spec.js
@@ -27,7 +27,10 @@ describe('reporter', () => {
   const mockHelper = {
     isDefined: (v) => helper.isDefined(v),
     merge: (...arg) => helper.merge(...arg),
-    mkdirIfNotExists: mkdirIfNotExistsStub,
+    mkdirIfNotExists: (dir, done) => {
+      mkdirIfNotExistsStub(dir, done)
+      setTimeout(done, 0)
+    },
     normalizeWinPath: (path) => helper.normalizeWinPath(path)
   }
 
@@ -166,7 +169,6 @@ describe('reporter', () => {
       const dir = rootConfig.coverageReporter.dir
       expect(mkdirIfNotExistsStub.getCall(0).args[0]).to.deep.equal(resolve('/base', dir, fakeChrome.name))
       expect(mkdirIfNotExistsStub.getCall(1).args[0]).to.deep.equal(resolve('/base', dir, fakeOpera.name))
-      mkdirIfNotExistsStub.getCall(0).args[1]()
       expect(reportCreateStub).to.have.been.called
       expect(mockPackageSummary.execute).to.have.been.called
       const createArgs = reportCreateStub.getCall(0).args
@@ -192,7 +194,6 @@ describe('reporter', () => {
       const subdir = customConfig.coverageReporter.subdir
       expect(mkdirIfNotExistsStub.getCall(0).args[0]).to.deep.equal(resolve('/base', dir, subdir))
       expect(mkdirIfNotExistsStub.getCall(1).args[0]).to.deep.equal(resolve('/base', dir, subdir))
-      mkdirIfNotExistsStub.getCall(0).args[1]()
       expect(reportCreateStub).to.have.been.called
       expect(mockPackageSummary.execute).to.have.been.called
     })
@@ -213,7 +214,6 @@ describe('reporter', () => {
       const dir = customConfig.coverageReporter.dir
       expect(mkdirIfNotExistsStub.getCall(0).args[0]).to.deep.equal(resolve('/base', dir, 'chrome'))
       expect(mkdirIfNotExistsStub.getCall(1).args[0]).to.deep.equal(resolve('/base', dir, 'opera'))
-      mkdirIfNotExistsStub.getCall(0).args[1]()
       expect(reportCreateStub).to.have.been.called
       expect(mockPackageSummary.execute).to.have.been.called
     })
@@ -246,7 +246,6 @@ describe('reporter', () => {
       expect(mkdirIfNotExistsStub.getCall(1).args[0]).to.deep.equal(resolve('/base', 'reporter1', 'opera'))
       expect(mkdirIfNotExistsStub.getCall(2).args[0]).to.deep.equal(resolve('/base', 'reporter2', 'CHROME'))
       expect(mkdirIfNotExistsStub.getCall(3).args[0]).to.deep.equal(resolve('/base', 'reporter2', 'OPERA'))
-      mkdirIfNotExistsStub.getCall(0).args[1]()
       expect(reportCreateStub).to.have.been.called
       expect(mockPackageSummary.execute).to.have.been.called
     })
@@ -277,7 +276,6 @@ describe('reporter', () => {
       expect(mkdirIfNotExistsStub.getCall(1).args[0]).to.deep.equal(resolve('/base', 'reporter1', 'defaultsubdir'))
       expect(mkdirIfNotExistsStub.getCall(2).args[0]).to.deep.equal(resolve('/base', 'defaultdir', 'CHROME'))
       expect(mkdirIfNotExistsStub.getCall(3).args[0]).to.deep.equal(resolve('/base', 'defaultdir', 'OPERA'))
-      mkdirIfNotExistsStub.getCall(0).args[1]()
       expect(reportCreateStub).to.have.been.called
       expect(mockPackageSummary.execute).to.have.been.called
     })
@@ -585,6 +583,38 @@ describe('reporter', () => {
       expect(log.error).to.not.have.been.called
       expect(log.warn).to.have.been.called
       expect(done.calledOnce).to.be.true
+    })
+
+    it('should handle unexpected errors during the coverage generation', async () => {
+      const errorSpy = sinon.spy()
+      const doneSpy = sinon.spy()
+
+      const customLogger = {
+        create: () => {
+          return {
+            debug () {},
+            info () {},
+            warn () {},
+            error: errorSpy
+          }
+        }
+      }
+
+      const error = new Error('Directory creation failed!')
+
+      const customHelper = {
+        ...mockHelper,
+        mkdirIfNotExists: async (_, done) => done(error)
+      }
+
+      reporter = new m.CoverageReporter(rootConfig, customHelper, customLogger)
+      reporter.onRunStart()
+      browsers.forEach(b => reporter.onBrowserStart(b))
+      reporter.onRunComplete(browsers)
+      await reporter.onExit(doneSpy)
+
+      expect(errorSpy).to.have.been.calledOnceWith('Unexpected error while generating coverage report.\n', error)
+      expect(doneSpy).to.have.been.calledOnceWith(1)
     })
   })
 })